+++ components/account/AccountPage.tsx
@@ -51,73 +51,49 @@ const useStyles = makeStyles((theme: Theme) =>
   })
 );
 
-const AccountPage = () => {
-  const router = useRouter();
 
-  //if (!router.query.account) return <ContentLoader />;
-  //console.log(router.query.block);
-  const address = router.query.account;
-  console.log(address);
-  const { loading, error, data } = useQuery(GET_ACCOUNT_DETAILS, {
-    variables: { address },
-  });
+
+const AccountPage = (props: any) => {
   const classes = useStyles();
-  if (loading) return <ComponentLoader />
-  if (error) return <ErrorMessage message={error.message} />
+
+  const accountAddress = props.address
 
   return (<>
     <Grid container className={classes.root}>
-      <Hidden lgUp>
-        <Grid item xs={12} lg={5} className={classes.bottomPadding}>
-          <AddressCard address={data.account.address} />
-        </Grid>
-      </Hidden>
 
-      <Hidden lgUp>
-        <Grid item xs={12} lg={5} className={classes.bottomPadding}>
-          <AccountOverview balance={data.account.balance} />
-        </Grid>
-      </Hidden>
+      <Grid item xs={12} md={8} className={classes.bottomPadding}>
+        <AddressCard address={accountAddress} />
+      </Grid>
 
-      <Hidden mdDown>
-        <Grid item xs={12} lg={5} className={classes.bottomPadding}>
-          <AddressCard address={data.account.address} />
-          <p></p>
-          <AccountOverview balance={data.account.balance} />
-        </Grid>
-      </Hidden>
 
-      <Hidden mdDown>
-        <Grid item xs={12} lg={5} className={classes.bottomPadding}>
-          <AccountDetails />
-        </Grid>
-      </Hidden>
 
-      <Grid item xs={12} lg={5} className={classes.bottomPadding}>
+      <Grid item xs={12} md={8} className={classes.bottomPadding}>
+        <AccountOverview address={accountAddress} />
+      </Grid>
+
+
+      <Grid item xs={12} md={8} className={classes.bottomPadding}>
         <AccountTransactions />
       </Grid>
 
-      <Grid item xs={12} lg={5} className={classes.bottomPadding}>
-        <InternalTransactions />
-      </Grid>
 
-      <Grid item xs={12} lg={5} className={classes.bottomPadding}>
+      <Grid item xs={12} md={8} className={classes.bottomPadding}>
         <CoinBalanceHistory />
       </Grid>
 
-      <Grid item xs={12} lg={5} className={classes.bottomPadding}>
+      <Grid item xs={12} md={8} className={classes.bottomPadding}>
         <Downtime />
       </Grid>
 
-      <Grid item xs={12} lg={5} className={classes.bottomPadding}>
+      <Grid item xs={12} md={8} className={classes.bottomPadding}>
         <ValidatedBlocks />
       </Grid>
 
-      <Hidden lgUp>
-        <Grid item xs={12} lg={5} className={classes.bottomPadding}>
-          <AccountDetails />
-        </Grid>
-      </Hidden>
+
+      <Grid item xs={12} md={8} className={classes.bottomPadding}>
+        <AccountDetails />
+      </Grid>
+
     </Grid>
   </>
   );