+++ pages/account/[account].tsx
@@ -13,25 +13,20 @@ const useStyles = makeStyles((theme: Theme) =>
       paddingTop: "2%",
     },
 
-    bottomPadding: {
-      overflow: "auto",
-      // paddingTop: '1.5%',
-      // paddingBottom: '1.5%',
-      //   padding: "1rem",
-    },
+
   })
 );
 
 export default function Account() {
   const classes = useStyles();
   const router = useRouter();
-  const { Account } = router.query;
+  const accountAddress = router.query.account
 
   return (
     <Layout>
       <Grid container className={classes.root}>
-        <Grid item xs={12} sm={10} className={classes.bottomPadding}>
-          <AccountPage />
+        <Grid item xs={12} sm={10}>
+          <AccountPage address={accountAddress} />
         </Grid>
       </Grid>
     </Layout>